Update test paths
